Make ring request :headers keys be lowercase strings (#7)
From https://github.com/ring-clojure/ring/wiki/Concepts#requests:

`:headers` A Clojure map of lowercase header name strings to corresponding
header value strings.

Without this change, the `:headers` map is not in a format that other
ring middlewares expect. For example, `wrap-json-params` which is used
in the README of this middleware expects there to be `content-type`
header, whereas with the same example keywordizing lambda invocation
payload and the current version of ring-apigw-lambda-proxy the `:headers`
map will end up containing key `:Content-Type`.